Goalkeeper Beta version with predict

+++ Code/Mother board/code/src/libs/VectorToMove/VTM.h
@@ -1,8 +1,8 @@
 struct VectorToMove
 {
-    int _angle;
-    int _mod;
-    VectorToMove(int angle, int mod)
+    double _angle;
+    double _mod;
+    VectorToMove(double angle, double mod)
     {
         _angle = angle;
         _mod = mod;
